docs: add JSDoc comments to sui-common
