+++ packages/sui-common/.stokedrc.json
@@ -1,0 +1,11 @@
+{
+  "version": "1.0.0",
+  "runs": [
+    {
+      "timestamp": "2025-04-05T13:14:26.887Z",
+      "jsDocBlocksAdded": 96,
+      "componentsDocumented": 10,
+      "testMode": false
+    }
+  ]
+}
