+++ packages/sui-common/src/Colors/index.ts
@@ -1,1 +1,3 @@
-export * from './colors';
+/**
+ * Module for exporting color-related utilities
+ */
