+++ packages/sui-common/src/Ids/index.ts
@@ -1,6 +1,10 @@
 import namedId from './namedId';
 import useIncId from './useIncId';
 
+/**
+ * Exported functions for generating named IDs and incrementing IDs.
+ * @module idGenerators
+ */
 export { useIncId, namedId };
 export * from './namedId';
 export * from './useIncId';
