+++ packages/sui-common/src/Ids/namedId/index.ts
@@ -1,3 +1,8 @@
+/**
+ * Exported namedId module.
+ * @module namedId
+ */ 
+
 import namedId from './namedId';
 
 export default namedId;
