+++ packages/sui-common/src/Ids/useIncId/index.ts
@@ -1,3 +1,8 @@
+/**
+ * Custom hook to generate an incremental ID.
+ * 
+ * @returns {number} - The generated incremental ID.
+ */
 import useIncId from './useIncId';
 
 export * from './useIncId';
