+++ packages/sui-common/src/LocalDb/index.ts
@@ -1,5 +1,11 @@
+/**
+ * Imports LocalDb from './LocalDb' and VideoDb from './VideoDb'.
+ * Exports LocalDb as default and re-exports LocalDb and VideoDb.
+ */
 import LocalDb from './LocalDb';
 import VideoDb from './VideoDb';
+
 export default LocalDb;
 export * from './LocalDb';
 export * from './VideoDb';
+*/
