+++ packages/sui-common/src/MimeType/index.ts
@@ -1,3 +1,7 @@
+/**
+ * Exporting necessary modules related to mime types
+ */
 export * from './MimeType';
 export * from './IMimeType';
 export * from './StokedUiMime';
+*/
