+++ packages/sui-common/src/index.ts
@@ -1,3 +1,8 @@
+/**
+ * @packageDocumentation
+ * This package provides various utility modules and components for handling data, colors, state management, types, IDs, fetching with backoff, local database operations, MIME types, and data loading using Grok.
+ */
+
 import LocalDb from './LocalDb';
 import GrokLoader from "./GrokLoader/GrokLoader";
 
@@ -10,4 +15,3 @@ export * from './FetchBackoff';
 export * from './LocalDb';
 export * from './MimeType';
 export { GrokLoader };
-
